Create EDGAR functions and summary programs
The edgar_functions program stores the defined functions used to
interact with the SEC EDGAR API. The summary-of-edgar-data.py provides
high level summary information for a company ticker entered. Was
initially created as a test of the edgar_functions program but can be
expanded later to pull more summary information for the ticker entered.